requirements.txt
psycopg2-binary
python-dotenv
pymongo[srv]
redis
tqdm